requirements.txt
asyncpg==0.31.0
python-dotenv==1.2.1
Telethon==1.42.0
pika==1.3.2